added operation class and updated broadcast class
